Remove noisy navigation description
This commit removes about 10 lines of perfectly accurate
description of how to navigate the table of contents, but there is no
reason for this! The TOC is easy to navigate as is, the huge block of
text looks scary and people probably won't read it anyway.
I just added a single line of italic text above TOC to stress it is
hyperlinked. And that's probably all a reader needs.